Take a less naive approach to async SQLite
My (currently, pretty limited) understanding here is that SQLite is
essentially single threaded. That's a bit of a simplification since
SQLite has different threading modes, including a multi-thread mode,
but even with that mode there are a bunch of limitations and the db
essentially "wants" to function single threaded. This makes sense for
a db like SQLite that is essentially just doing local file IO. The
underlying file IO isn't async and the db is constrained by that.

This poses something of a challenge when using SQLite in the context
of Axum, since Axum is multi-threaded and you can have situations
where multiple threads all want to interact with the database.

Based on my research (i.e. random internet searching) there are a few
ways to deal with this mismatch.

1. Share a single connection to the database across threads by
   guarding the connection with a mutex. Since threads will
   potentially block waiting to acquire the mutex lock in this case,
   you might also ensure all database operations are run on a thread
   pool dedicated to running blocking functions, for instance via
   `tokio::spawn_blocking`.

2. Establish a connection pool with multiple db connections, for
   instance, via the `deadpool` crate. Each thread must attempt to get
   a connection from the pool and will block and wait (up to a timeout
   period) if the pool is fully drawn. I think that, since the
   underlying db is still essentially single threaded, this more or
   less pushes the question of how to manage work from competing
   connections down to the db.

3. Spawn a task/thread specifically dedicated to all interaction with
   the database, and use message passing (i.e. via channels) to
   interact with it across threads.

Prior to this commit, we were sort of naively attempting strategy 1
(though in a way that I believe was incorrect). This commit changes
things so that now we use strategy 3, via the `tokio-rusqlite` crate.