Move db queries from controllers to models
This commit moves code that reads from or writes to the database from
the controller layer to the model layer of the app. It does this by
defining methods that execute queries onto `Paste` (currently the only
extant model). The database connection is provided to the model via
dependency injection - each query function takes an argument of the form
`db: &Database`.

While I am very confident that getting the database logic out of the
controller layer is the correct design decision, there were various
possible ways to achieve that. I considered alternative designs
including:

- a "driver" or "client" struct, potentially per model, that would then
  hang on the `Database` struct, and use the struct's `Connection` to
  execute queries.
- a distinct "queries" module containing all querying functions.
  Potentially with a more complicated series of functions for building a
  sql statement, binding or executing those statements (via a dependency
  injected db connection), and functions for specifying whether to
  marshall all, one, or an optional result. This would be pretty similar
  to (and was inspired by) the generated code produced by the
  `cornucopia` crate.

Ultimately, I decided these alternatives were overcomplicating, at least
for now. Having database interaction in close proximity to the models
seems logical to me. And using dependency injection to get the database
connection to querying functions, while somewhat repetitive, has the
advantage of being extremely explicit about which model methods are
going to act on the database, and which will not.