Add the MadimiOne display font
The plan is to use this a our logo font and maybe also for some headers.

The font is licensed under the Open Font License and has been subset for
only Latin characters. We are self-hosting it, since that is more
respectful of people's privacy than loading via someone else's CDN.