Add a title field to pastes
And wire it through to the templates. Also, realized along the way that
there was a dead code path in `pastes_show.html` that was created when I
improved 404 handling earlier, so this removes that as well.

The idea with this title field is that, eventually, it will:
- Allow `Paste.description` to be optional
- Serve as a way to infer file naming for a "download-paste-as-file"
  feature

But neither of those things are implemented at this stage, so for now
it's just another string field.