Add serde_rusqlite
This brings convenient serialization to and from the database.

But...figuring out how to handle potential serialization errors when
reading from the database was an issue here and I'm still not confident
I've landed on the best solution.

The core problem is that:

- `serde_rusqlite::from_rows` needs to be called on `rusqlite::Rows` and
  that means it must be called within the closure passed to
  `tokio_rusqlite::Connection::call`, because `Rows` can't be passed
  outside of that closure (I think because `Rows` doesn't implement
  Send?)

- But the closure for `tokio_rusqlite::Connection::call` also must
  return a Result with trait bounds `Send + 'static` and that doesn't
  seem to work with `serde_rusqlite::Error`.

The solution I landed on was to map `serde_rusqlite::Error` to
`tokio_rusqlite::Error::Other` which seems to have been included on
their error enum for this purpose (see
https://github.com/programatik29/tokio-rusqlite/issues/18). This works,
but it also requires dynamic dispatch, and "downcasts" (I think that's
the term?) `serde_rusqlite::Error` in a way that's different from how
error handling throughout the rest of the app works.