Modestly improve error logging
This makes it so that errors will at least get logged in a somewhat
informative manner, whereas before we'd just log that the server
returned a 500.

Also, removed the /errors/:error_code routes since I couldn't find a way
to isolate them only to dev environments that I was happy with. And
anyway, it's easy enough to generate a 500 if we need to see what
`views/500.html` looks like anyway.