Introduce TestApp struct
Longer term, the idea will be to hang supporting functionality that is
useful to a majority of integration tests on this.

But for now, it gives us access to the database in integration tests.
This access should be used sparingly, since it would be better for tests
to call an interface, like an api, that accesses state in a way that is
agnostic to the underlying implementation details. But that sort of
interface isn't always available or something that the app should be
exposing.

This also saves us from having to modify every test whenever
`common::spawn_app` changes.