Implement and test `GET /api/pastes`
This is the first of several JSON api endpoints I'm planning to build.
While primary interaction with `gluestick` will likely be via the web
app (and therefore, via the "html" api), it seems like a good idea to
also offer a JSON api for programmatic use cases.

The JSON api will also be useful to my integration tests, since I will
be able to call it to setup test states, rather than having to
interacting directly with the database. That's more ideal, since ideally
integration testing would treat the app as completely opaque and
wouldn't require knowledge of the database (which is an implementation
detail).

This commit also includes substantial reorganization of the integration
tests including:

- introduction of a helper `TestPaste` struct, to make serialization and
  deserialization of pastes in a testing context less of a pain

- introduction of `tests/main.rs` and reorganizing existing modules
  under it. This makes it so that cargo builds just one integration
  testing crate, rather than a new crate per test file.

- division of test cases into `html_api` and `json_api` modules, for
  purely organizational reasons