Clean-up controllers::Error
The basic idea here is that the controller layer doesn't really need to
enumerate every possible failure mode of all the layers below it. Many
of those failure modes are just 500 Internal Server Error from the
perspective of a user (and control flow) anyway.

Still haven't decided how to I want to render 400 and 401 errors to end
users, so those render blank for now.