Clean up json api error handling
This defines a dedicated error type for the json api controllers, which
is very similar to extant html api controller error, except that it
renders to json, rather than html.

Also was unhappy with how the 404 fallback and healthcheck routes were
organized in the codebase so did some renaming and moving around there.