Reorganize auth code
This gets rid of some dead code in `session.rs` by moving `SessionToken`
into that file.

It also creates an `extractors` folder, which may or may not turn out to
be a useful organization in the future. For now, it houses the auth
extractor that gets a user from a session token.