Extend paste creation to support visibility levels
Currently this supports "public" and "secret". Neither is truly "private"
in the sense that one must pass an auth check to view the paste. I may
add a third option for that in the future.

Also, I'm undecided on whether it should be possible to change a paste
from public to secret or vice versa, after it has been created. For now,
I've excluded this from the edit/update paths. The argument against is
that, at least for public -> secret edits, there's no real way to
ensure that the url for the paste wasn't scraped or otherwise recorded
while it was public, so it might give people the false impression of
more secrecy than we can guarantee.