Implement copy action for the show paste page
This is extremely basic. It does not provide any indication of success
or failure. It also does not attempt to ascertain whether the user's
browser supports the clipboard API, nor does it attempt to disable or
hide the copy action button in unsupported cases.

But it's simple, should work most of the time, and I'm unsure that the
marginal improvement from handling such things is worth the increased
complexity.