Add syntax highlighting of pastes
I suspect this will need to be optimized since it currently loads the
syntax highlighting theme and renders each paste body to highlighted
html on every read, which is pretty CPU ineffecient.

Solution to that would be to cache the highlighted html when a paste is
written or via a background job. But I'm not going to bother
implementing that until I've profiled and know for sure it's an issue.