Add a timeout layer with 10s timeouts
10s is a bit of an arbitrary number. But, current load testing against a
cheap VPS has the app hitting a requests per second plateau at a point
where p95 response times are still only 1.2s. So, I think 10s is
generous enough and that should hopefully become more true as I optimize
various things.

Also, `TimeoutLayer` returns a 408 status, not a 504. I'm sort of unsure
whether that's appropriate or not, but it does mean that browsers will
retry the request. Firefox, for instance, retries 10 times. Just
something to be aware of when testing this.