Switch to cursor-based pagination
As noted in the prior commit, this is substantially more complicated
than offset-based pagination. But in exchange for that complexity we get
better query performance from any point in the pagination.

Ultimately going with this since I feel I was mostly able to get the
implementation to a point where the additional complexity isn't too
burdensome.