Refactor syntax highlighting logic
This commit pulls syntax highlighting logic, including all
of the caching logic, out of the `Paste` model and
controllers, and centralizes it into a helper module.

Additionally, the logic that attempts to fetch a cached verison of the
syntax highlighting prior to generation has been improved so that
various race condition bugs are now impossible.