Add basic user tests, trying new test patterns
The new tests in this commit uncovered a bug that I will fix in a
following commit.

The new test patterns trialed in this commit include:

- Using the builder pattern for setting up test state
- Random generation of some test state
- An Error and Result type specific to testing, via a prelude module

My goal going in was to make test state more concise, easier to read,
and easier to compose. I think the changes mostly succeed at that goal.
Further improvements that could be made in the future include:

- Rolling these patterns out to the rest of the test suite, especially
  for the pastes tests
- Additional refactoring on the Client and TestApp structs to further
  reduce verbosity
- Using the prelude pattern elsewhere, potentially for the models and
  controllers modules