Add more signup tests and fix bugs
The additional testing uncovered two bugs, both of which are fixed in
this commit.

The first bug was that we were allowing people to signup with
zero-length passwords. That is now fixed and we enforce a length
requirement of min 8, max 256 on passwords at signup.

The second bug was that the app's user email validation was actually
less strict than the database constraint on users.email. I thought it
was the other way around, but apparently the html5 validation (which the
app uses) doesn't require a .tld on emails. This is also now fixed - the
db now only requires %_@_%.