Try out a pattern for form verification
Verification, as opposed to Validation - where the distinction is that
verification is checking input against state like the database, while
Validation  is a check that can be performed purely by inspecting the
input itself.

This is still pretty rough and needs further refinement, especially
around error handling. There are unwrap calls in some of the verify
methods that will need to be handled, and I'd like to tie this in to
controller error handling so that it is less verbose.