Add `watch-testless` to Justfile
This is the same as the `just watch` command, except it does not run the
test suite on each file change.

This is useful in situations where you want to repeatedly rebuild the
app, but also would prefer to iterate quickly. A partial recompile
usually takes less than a second, whereas the test suite takes 5-6
seconds.

The specific use-case I have for this is html editting. The app's html
is built using the `askama` crate's templating engine, which requires a
recompile on every change. But I don't need to re-check the tests on
every little change, and doing so slows down my workflow substantially.
So, this provides a useful alternative for that scenario.