Rebaseline the db schema
This also eliminates the `ON DELETE CASCADE` foreign key callbacks from
tables that do not represent fundamentally transient data (so,
everything that isn't either a session or a cache).

I think those callbacks are a footgun on tables with data that you
actually care about, since they will trigger if you drop the table
referenced by the foreign key, potentially cascading and deleting most
of your data. That may seem like a hypothetical scenario, but SQLite
doesn't support changing constraints via `ALTER TABLE` so it's
relatively common in migrations that need to change a constraint to copy
a table over to somewhere new, drop the original, and then rename the
new table back to the original. If you do that with a bunch of foreign
keys referencing the dropped table that have set `ON DELETE CASCADE` you
will delete a lot of data you didn't intend to delete...