Use avif images
Avif reduces the file size of our images by ~80%, which is too good of a
savings to pass up.

It isn't supported by IE11 or somewhat older versions of Safari, but
HTMX dropped IE11 support so we're already incompatible with that, and
we can use `<picture>` to fallback to the png versions of the images on
older Safari versions.