Add progressive image loading
This idea here is to load a tiny pixelated version of each image
quickly, and then load the full images lazily, swapping them in for the
pixelated versions on load.

This has a few trade-offs:
- on extremely slow connections, images are previewed in pixelated
  form substantially earlier than when they would load otherwise
- but we have to perform an extra ~1-2kb request per image to fetch the
  pixelated previews
- and we have to use javascript to swap out the real image for the
  preview

It's a pretty substantial increase in complexity, especially since we
have to provide both png and avif versions of the previews, in case a
user's browser doesn't have avif support yet. But on really slow
connections its a pretty nice quality of life improvement, so I think
I'm willing to take the complexity hit for now.