Add a background task that expires idle sessions
This will expire any session that is idle for more than 8 hours.

In order to achieve that functionality, it adds an `updated_at` field to
the `sessions` table and a method that "touches" or updates that field
every time a user is authenticated via the session. Since we don't
currently have a prod database that we care about, this is done WITHOUT
a proper database migration.