Refactor `Session` model
This changes `Session` so that it is now essentially a `User` and
`SessionToken` pair. And `SessionToken` is now a proper model that aims
to be the app-side representation of an entry in the `session_tokens`
database table. The old version of `SessionToken` and
`HashedSessionToken` have been renamed to `UnhashedToken` and
`HashedToken` in order to avoid the namespace collision with the new
`SessionToken` struct.

Why do this? I wanted a struct that more closely models the
`session_tokens` database table, because having that provides a good
place for functionality that needs to query or manipulate that table.
This isn't strictly needed for `Session`, but I think it will be quite
desirable for `ApiSession` where the concept of api keys that are
managed, deleted, rolled, etc. is going to be very prevalent. And since
I want `Session` and `ApiSession` to mostly be symetric in their design,
that meant refactoring `Session` like this as well.