Have htmx render error responses
By default, HTMX does not render non-200|300 responses. This changes
that behavior so that the response body on 400s and 500s is swapped in
within the `<body>` tag, similar to how 200s are handled by
`hx-boosted`.

Since our error responses include a rendered error template, this new
behavior better matches what we want to happen in the event that an
unexpected error occurs.