Allow limited editting of paste visibility
This allows secret pastes to be made visible, but public pastes are
permantently public and cannot be udpated back the other direction.

The logic behind that is that people can opt to disclose a paste that
was previously secret, but they can't make an already-public-paste
secret since the paste may have been indexed or seen by someone.