Format html with prettier and its jinja template plugin
CSS was already being formatted with prettier, but I had been using
djlint to format Askama templates prior to this commit.

Askama, of course, does not follow the exact same format and syntax as
Jinja or Django Template, but it is pretty close. I am also not aware of
any code formatter for Askama, beggars can't be choosers, and etc.

Why prettier over djlint? A few reasons:
- djlint has an open call for maintainers and I'm worried it's going to
  get orphaned
- djlint requires python, which I otherwise don't really use for
  anything.
- I already use prettier for formatting several other markup and coding
  languages.